update: move page styles to pages/styles
