+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Layout } from '../layout/Layout'
 import Seo from '../components/Seo'
-import '../styles/404.css'
+import './styles/404.css'
 
 const NotFoundPage = () => (
   <Layout location="Oops!">
