+++ src/pages/cv.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Seo from '../components/Seo'
 import { Layout } from '../layout/Layout'
 import { ComingSoon } from '../layout/ComingSoon'
-import '../styles/cv.css'
+import './styles/cv.css'
 
 const CvPage = () => (
   <Layout location="CV">
