+++ src/pages/index.tsx
@@ -6,7 +6,7 @@ import { Tab } from '@headlessui/react'
 import { Layout } from '../layout/Layout'
 import { ArrowNarrowRightIcon } from '@heroicons/react/outline'
 import { Gallery, About, Skills, Resume } from '../components/hero'
-import '../styles/index.css'
+import './styles/index.css'
 
 const IndexPage = () => {
   const [focusRing] = React.useState(false)
