+++ src/pages/me.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Seo from '../components/Seo'
 import { Layout } from '../layout/Layout'
 import { ComingSoon } from '../layout/ComingSoon'
-import '../styles/me.css'
+import './styles/me.css'
 
 const MePage = () => {
   return (
