+++ src/pages/styles/portfolio.css
@@ -1,4 +1,3 @@
-/* Portfolio Page */
 .portfolio {
   @apply container mx-auto flex-1 flex-col justify-between space-y-6 px-4 py-4 2xl:px-0;
 }
@@ -15,7 +14,6 @@
   @apply grid grid-cols-1 gap-6;
 }
 
-/* Portfolio Entry Card */
 .portfolio-entry {
   @apply flex w-full flex-col space-y-3 rounded-xl bg-lightest p-3 shadow-inner-md dark:bg-white/5 lg:flex-row lg:space-y-0;
 }
@@ -47,7 +45,7 @@
 }
 
 .portfolio-entry > section header .title {
-  @apply text-xl uppercase tracking-wide font-semibold text-primary dark:text-tertiary;
+  @apply text-xl font-semibold uppercase tracking-wide text-primary dark:text-tertiary;
 }
 
 .portfolio-entry > section header .dates {
