+++ src/pages/{MarkdownRemark.frontmatter__slug}.tsx
@@ -3,6 +3,7 @@ import { graphql } from 'gatsby'
 import { Layout } from '../layout/Layout'
 import { defineCustomElements as deckDeckGoHighlightElement } from '@deckdeckgo/highlight-code/dist/loader'
 import { GatsbyImage, getImage } from 'gatsby-plugin-image'
+import './styles/blogpost.css'
 
 export default function Template({ data }) {
   const { markdownRemark } = data
